fix #10496 OOS when using DSU (#10497)
the old code tried to prevent the creation of a dummy undo stack entry for toggling shroud updates if there are no elements on the undo stack before. However the old way of doing this resulted in the action being removed from the replay recorder aswell, leading to OOS.

Now we just always block undo in this case.

The optional paramter in get_auto_shroud does in theory allow us to make it undoable again later in a save way.