Revert 'Balancing Update' to Dwarvish Miner
Resets HP and XP back to their values from 10 months ago.

hitpoints: 25 -> 22
experience: 20 -> 24

The Dwarvish Miner is already a tough unit with great resistances, movement_type, traits, advancements and damage. Giving it almost the same HP as the Wose Sapling is out of proportion, both in relation to the HP of comparable units, and because it makes the Miner impossible to kill by units of the same level.

Also, setting the XP to 20 means an intelligent Miner can advance with two L1 kills, which is too fast for a unit that advances into a Dwarvish Guardsman/Fighter. Those two are among the toughest default L1 recruits, so this gives a massive advantage to the side that can acquire them so *quickly*.