Format CONTRIBUTING.md's example code with clang-format
This converts spaces to tabs in the example, but it's not just
whitespace changes.

The #include "gettext.hpp" and #include "help/help.hpp" lines are now
sorted alphabetically.

The lambda function is short, thus all on a single line.

Add a note that we're incrementally moving towards the standard rather
than doing a project-wide reformat that would break the git-blame
history.