Test scenario: Give the Shyde a new slows that only works in forest
Weapon specials can have location filters, even if they're applied
directly to the weapon rather than being used like abilities. This
gives the Shyde a special which is inactive except on forest hexes.

Currently (commit ee7f4c047a4) the sidebar always shows it as active,
because the sidebar uses a weapon context that doesn't know which
hex it should be used in.

Merging as a test tool, without defining what the correct behavior is.
I think the sidebar should show it as inactive except when moving to
or attacking from a forest hex, however that needs discussion about
whether other abilities are consistent, and the cases involving hexes
under shroud or fog need checking to ensure that the sidebar doesn't
leak information.