Add a 'real_cumulative' attribute that acts like 'cumulative' in Leadership, but for other abilities and all numerical attributes.
Because of the double meaning of 'cumulative' (cumulative value for Leadership, but choosing the highest value in other cases), to have a truly cumulative attribute, I need to use a different name.

If this pull request and https://github.com/wesnoth/wesnoth/pull/10646 are merged, the 'accuracy' and 'parry' attributes will become obsolete and can be replaced by 'chance_to_hit'. Other abilities or special weapons could also benefit from this new attribute.