Add [clear_chat], [store_zoom], and optional scrolling support
Add a couple simple (but potentially useful) wml actions, and add support for scrolling only if a tile is not currently visible.

[clear_chat]: as the name implies, clears recent chat messages. Could be used with cutscenes or transitions between different phases of a multiplayer scenario, or also if a scenario is using chat messages to provide information to the player.

[store_zoom]: implements request in #7423. Useful if a cutscene or other animation requires a fixed (non-relative) zoom level, so it can be set back, and I could also envision need for animations to scale based on current zoom level.

In addition, extends [scroll_to] and [scroll_to_unit] to support only_if_needed; if this is true then the map will only scroll if the hex or unit is not already visible. Primarily intended for extended dialog scenes with animations where units are (probably) all visible anyway so they can be less choppy.

Finally, extends [kill] with a scroll attribute. Defaults to true. If set to false, then map will not scroll to the unit being killed. Similar use cases as only_if_needed.