Remove confusing locale dir differences
All variants are either relative to the data dir or not. Relative to the
current working directory is just what happens naturally if LOCALEDIR is a
relative path.

To keep the same behavior on windows make sure HAS_RELATIVE_LOCALEDIR is set.
CMake does that on windows but SCons does not.

As far as I can tell USE_INTERNAL_DATA is not used and unless
HAS_RELATIVE_LOCALEDIR is also set for some reason the behavior stays the same
except for a bug when the CWD cannot be determined. If get_cwd has an error it
returns the empty string so in that case a "/" would be prepended which is just
wrong.