Timers for halo updates fixed
The timers for halo updating where off before, this caused two issues. 1: The halo would often be stuck on the last frame in the cycle. 2: Since smaller parts of the halo are also updated by other animation update functions there could be glitches where parts of the halo are 1 frame behind the rest.