Fix unit moving frame updates
We now make a list of the hexes that a frame is on. On the new frame we add that list (last_redraw_hexes_) to tiles that needs redrawing.

This solves the issue and should be more future proof.

I also commented out a lot of lines that currently does not seem to do anything, I need a second opinion before I remove any.

hm