attack_type: mark describe_modification as static
The generated description relied only on the modification's config and not on any object state. This made its one usecase confusing (and redundant), since the same output was produced regardless of the attack through  which this function was invoked. This was also inefficient, since the same modification description was generated for every attack object, and confusing, since it appeared as if subsequent descriptions were overwriting prior ones.