animation wml: fix range padding syntax (#10627)
* animation wml: fix range padding syntax

Previously it padded the number to the number of leading zeroes plus one, now it pads to number of digits of the number, so for example to get the range 010, ... , 090, one previously  would have to write `[0010~0090]` now one needs to write `[010~090]`.

The error message below also seemed to already expect the new syntax.

In the most common cases like [002~100] the result is the same, so i don't expect breakage.