Use a struct + shared_ptr for untis abilities.
Currently the struct is only a plain of tagname+config, bit the idae is to turn it into a more "normal" class at some point. in particular:

    we can store/cache filter objects speeding up ability evaluation
    we can store/cache theabilitiesproperties speeding up ability evaluation
    we can share the [event]s removing a lot of dublicate data.

Using shared_ptrs also makes it easier to make a the code more robust against bad wml/lua removing a units ability in a lua filter execution.