remove simple check parameter
This just makes refactoring of abilities.cpp simpler for now.

it also uncovered a bug in the backstab check in aspect_attacks.cpp:
it checks for the tag name but it should check for id, it seems like
at some point has_special was changed to check for tags only and this
line was overlooked.