Fix & Refactor help links in ability tooltips
Previously different parts of the code assumed different formats for help
topic id generation of abilities, breaking help links in ability tooltips.
weapon specials tooltips did not have this issue, but were also refactored
in this commit to behave the same.

This commit fixes that and while doing so refactors all the tooltip functions
to return a custom struct

f prev