Android: disable broken unpack status check
This particular check stopped updates from downloading when the version is changed. The unpacking should still work as usual since downloadFile() only downloads if file is newer on server, and unpacking only happens if the zip archive exists. Corrupt downloads are also handled inside downloadFile(), so the if check is redundant.
Also converted manual stream closing with try-with-resources.