Add the PO files for wesnoth-dod and wesnoth-h2tt
Many of the CI builds expect the .po files to already exist, and
fail without them.

For SCons builds, the .pot file needs to exist, because po/SConscript
uses the existing files to find out which textdomains it needs to
update.

Both were already added to CMakeLists.txt, so there are no CMake
changes in this commit.

The en@shaw.po files have been cleared of automatic translations. The
en_GB.po file contains the automatic copy of the American English text.