Add a [suppress_lower_priority] subtag to abilities and specials with value for suppress abilities/specials of lower priority who matches and extent priority to berserk, damage_type,plague and swarm
The priority system with priority= certainly allows you to change the order of calculations, for example, to have the base value multiplied before any other operation, but it doesn't allow you to exclude certain calculations. So, for example, if two [chance_to_hit] effects with add= affect the same fighter, one with a value of a and the other with a value of b, regardless of whether the priority is the same or not, chance_to_hit will always be base_value + a + b.

To complement this, I'm adding the [suppress_lower_priority] tag, which can not only suppress lower-priority abilities corresponding to the filter, much like overwrite_specials=[overwrite], but unlike the latter, it's not limited to special weapons, but also applies to abilities like `[heals], [regenerate], or [leadership], among others.