add helper foreach_active_special
previously the ability-related keys of weapon
filters were read in various functions in
abilities.cpp. Now they are again only read on in
the attack_type::matches_filter function,
So that we can hopefully at some point change
the weapon filter code
to have separate parse and evaluate steps,
just like the unit filter code currently has.

The original plan was to use this function for
attack_type::get_specials_and_abilities too but
there is a small problem: if the specials context
is not set i don't have the owner unit and the
active_ability_list needs the unit (or rather the
units location). The old code just falls back to
self_loc_ but that's usually just invalid too, so
it doesn't magically fix the problem either, so
i'm delaying that until i have a better solution.