Minor optimisations in sdl::mask_surface
Merge the loop conditions, and allow a variable to saturate bitwise
instead of having a branch to set a boolean.

Makes a slight improvement when compiling for release with Clang-22 on
x86_64 Linux, but for GCC-15 this roughly triples the speed (GCC-15's
code is still slower than Clang-22 here).

Avoiding branching in sdl::in_mask_surface gave a slight improvement
in both compilers, at least when testing with the images shown when
running `wesnoth -t`.