Ability/Weapon specials: support more attributes as po variables
This adds support for "value", "add", "sub", "multiply", "divide", "max_value" and "min_value", in addition to "type" for plague in #10293.
They could be accessed inside translatable strings in ability/weapon special definitions as $value, $add and so on.

As an example, UtBS weapon special "parrying" have been refactored to use this system.