Defaulted the constructors of blacklist_pattern_list and binary_paths_manager.  =default; means less code and less code means less to think about.
For blacklist_pattern_list, in member functions that take one or more expensive to copy container parameters, and that also take ownership of those parameters (ie, retain an internal copy), changed the way the parameters are passed from const& to by-value.  Updated these functions to std::move from the parameter internally, and updated callers to std::move in where possible.  In "take ownership" situations by-value is a more effecient and flexible approach than const&, because const& will always require that a copy be made.  By-value allows a user who no longer needs the value to avoid the copy by std::moving, while also accomodating users who *do* need to retain the value.  An alternative approach would be to define const& and && overloads of each of these member functions, but that's twice as much code and would only yield a very marginal performance benefit; a parameter being moved into a && overload (or a temporary being passed in) would experience one less move operation than it would in the by-value case, but moves are pretty cheap.