Change terrain_builder::set_terrain_rules_cfg() to clear building_rules_ by calling .clear() rather than by swapping it with an empty container.
set_terrain_rules_cfg, among other things, clears the contents of building_rules_,  a std::multiset.  It does this by swapping it with an empty multiset created on the stack, which is then destroyed when the method returns.  There is a confusing comment justifying this approach in the interest of avoiding "progressive memory degradation".  The author of the comment and the code likely had in mind containers like std::vector, for which clear() will set the size to 0 but will not actually release the memory, leaving the capacity() of the container unchanged.  std::multiset does not have the concept of capacity, and clear() will release the memory, so this confusing swap approach is not necessary.