Make terrain_type_data the source of truth
terrain_type_data is intended to be a singleton, but is passed around via a shared_ptr.  When terrain_type_data::reset reloads the singleton instance owned by terrain_type_data, anyone else holding a shared pointer obtained from terrain_type_data::get() is left with a pointer to a stale copy, or at least to a version of terrain_type_data different from what terrain_type_data::get() will subsequently hand out.  #10462 is caused by exactly this.  When the user presses F5 in the editor, the terrain.cfg file is re-read and terrain_type_data::singleton_ correctly updated with this new data, but gamemap::tdata_ is a shared pointer that keeps the old instance alive, and the terrain pallette in the editor is (confusingly) populated by calling methods on gamemap, which continues to hand out this stale data.

This change converts terrain_type_data::singleton_ to a unique_ptr, and makes terrain_type_data::get() the source of truth tor terrain_type_data.  I am also making ::get() return a reference rather than a pointer.  terrain_type_data is initialized very early on by the game_config_manager, so the null case is probably not a big concern.  In the longer term, it might be better to make terrain_type_data owned by the gcm, so that it can provide a guarantee that it will always be around when it is asked for.