Reintegrate terrain_type registry with game config manager (#10780)
Fixes #10462. Thanks to @vpbroadcastq for tracking down the lifetime issue. Their analysis is copied below:

terrain_type_data is intended to be a singleton, but is passed around via a shared_ptr. When terrain_type_data::reset reloads the singleton instance owned by terrain_type_data, anyone else holding a shared pointer obtained from terrain_type_data::get() is left with a pointer to a stale copy, or at least to a version of terrain_type_data different from what terrain_type_data::get() will subsequently hand out. #10462 is caused by exactly this. When the user presses F5 in the editor, the terrain.cfg file is re-read and terrain_type_data::singleton_ correctly updated with this new data, but gamemap::tdata_ is a shared pointer that keeps the old instance alive, and the terrain palette in the editor is (confusingly) populated by calling methods on gamemap, which continues to hand out this stale data.

This takes a slightly different approach than #10777. Though decoupling the terrain_type_data member from game_config_manager (ebdc9bb6a384f3ba25938bb23f479882f9612f2e) did not introduce the stale state bug, it at least made the code harder to understand by obfuscating ownership. The GCM is the only object that should manage the terrain type registry, so it doesn't make much sense to manage it through a self-owned singleton instance. Reintegrating with the GCM also means we can keep address of the cache consistent for the lifetime of the process, which means we don't need to update the pointers held by the gamemap. It turns out we don't actually need to recreate the whole terrain_type_data object when resetting it, since it uses a lazy initialization mechanism. All we need to is flag that initialization is needed and clear the stale registry.

It might be better to make the GCM hold a unique_ptr<terrain_type_data> to reduce header dependencies, but I might handle that separately. Other changes from #10777, including making gamemap hold a reference instead of a pointer or making the singleton getter return a reference, can also be done separately.