HttT S17: balance west vs south spawns (#10796)
HttT's S17 gives players the option to approach from both the west and the south. The options are meant to be similar, but I think the southern spawn has a harder time getting villages than the western spawn.

This PR attempts to both balance and differentiate the spawns. The southern spawn now has more villages, but they're harder to cap, while the western spawn has fewer villages, but easier to cap. This also makes the overall scenario slightly easier.