color_t: remove RGBA conversion functions
We use ARGB as our standard format everywhere, leaving the RGBA interface unused. In theory it's useful to have (it was after all split from the ARGB functions at one point), but it's better if we always use one byte pixel format everywhere.

The one main use for this besides the hash function (which is changed here) was the sorting operator for color_range, but I removed that as it didn't make much sense for color_range to be orderable in the first place.